implemented some constraints
sup constraint - tested
gw constraint - not tested

+++ src/main/java/geneticalg/timetable/constraints/Constraint.java
@@ -6,6 +6,6 @@ import geneticalg.timetable.entities.Course;
 
 public interface Constraint {
 
-	Integer checkConstraint(List<Course> timetable);
+	Long checkConstraint(List<Course> timetable);
 	
 }
